+++ plugins/turbo-plugin/scripts/pull-from-svn.ps1
@@ -6,7 +6,7 @@ param(
 Set-StrictMode -Version Latest
 $ErrorActionPreference = 'Stop'
 
-. (Join-Path $PSScriptRoot 'lib' 'common.ps1')
+. ([System.IO.Path]::Combine($PSScriptRoot, 'lib', 'common.ps1'))
 
 try {
     Probe-GitVersion